feat: better modules support
