--- rollup.config.js
@@ -1,4 +1,0 @@
-import umd from './rollup/umd';
-import cjsie11 from './rollup/cjs.ie11';
-
-export default [umd, cjsie11];
