--- src/utils/convertArrayToPathName.test.ts
@@ -1,11 +1,0 @@
-import convertArrayToPathName from './convertArrayToPathName';
-
-describe('convertArrayToPathName', () => {
-  it('should convert into path name', () => {
-    expect(convertArrayToPathName(['test', 2])).toEqual('test[2]');
-    expect(convertArrayToPathName(['test', 2, 'test'])).toEqual('test[2].test');
-    expect(convertArrayToPathName(['test', 2, 'test', 4, 'data'])).toEqual(
-      'test[2].test[4].data',
-    );
-  });
-});
