+++ superstruct/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "superstruct",
+  "amdName": "hookformResolversSuperstruct",
+  "version": "1.0.0",
+  "private": true,
+  "description": "React Hook Form validation resolver: superstruct",
+  "main": "dist/superstruct.js",
+  "module": "dist/superstruct.module.js",
+  "umd:main": "dist/superstruct.umd.js",
+  "source": "src/index.ts",
+  "types": "dist/index.d.ts",
+  "license": "MIT",
+  "peerDependencies": {
+    "react-hook-form": ">=6.6.0",
+    "@hookform/resolvers": "^1.0.0"
+  }
+}
