--- tsconfig.jest.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "strictNullChecks": true
-  },
-  "include": ["src/**/*.test.ts"]
-}
