+++ tsconfig.json
@@ -4,21 +4,16 @@
     "module": "ESNext",
     "target": "es2017",
     "moduleResolution": "node",
-    "outDir": "./dist",
-    "jsx": "react",
     "skipLibCheck": true,
-    "declaration": true,
-    "noEmit": false,
-    "allowSyntheticDefaultImports": true,
-    "lib": ["dom", "dom.iterable", "esnext"],
     "strict": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "noEmit": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "strictNullChecks": true,
     "isolatedModules": true
-  },
-  "include": ["src"],
-  "exclude": ["src/**/*.test.js", "src/**/*.test.ts", "src/index.ts"]
+  }
 }
