+++ vest/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "vest",
+  "amdName": "hookformResolversVest",
+  "version": "1.0.0",
+  "private": true,
+  "description": "React Hook Form validation resolver: vest",
+  "main": "dist/vest.js",
+  "module": "dist/vest.module.js",
+  "umd:main": "dist/vest.umd.js",
+  "source": "src/index.ts",
+  "types": "dist/index.d.ts",
+  "license": "MIT",
+  "peerDependencies": {
+    "react-hook-form": ">=6.6.0"
+  }
+}
