+++ yup/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "yup",
+  "amdName": "hookformResolversYup",
+  "version": "1.0.0",
+  "private": true,
+  "description": "React Hook Form validation resolver: yup",
+  "main": "dist/yup.js",
+  "module": "dist/yup.module.js",
+  "umd:main": "dist/yup.umd.js",
+  "source": "src/index.ts",
+  "types": "dist/index.d.ts",
+  "license": "MIT",
+  "peerDependencies": {
+    "react-hook-form": ">=6.6.0"
+  }
+}
